+++ src/arkreactor/VM/VM.cpp
@@ -536,9 +536,13 @@ namespace Ark
                 &&TARGET_LT_CONST_JUMP_IF_FALSE,
                 &&TARGET_LT_CONST_JUMP_IF_TRUE,
                 &&TARGET_LT_SYM_JUMP_IF_FALSE,
+                &&TARGET_GT_CONST_JUMP_IF_TRUE,
+                &&TARGET_GT_CONST_JUMP_IF_FALSE,
+                &&TARGET_GT_SYM_JUMP_IF_FALSE,
                 &&TARGET_EQ_CONST_JUMP_IF_TRUE,
                 &&TARGET_EQ_SYM_INDEX_JUMP_IF_TRUE,
                 &&TARGET_NEQ_CONST_JUMP_IF_TRUE,
+                &&TARGET_NEQ_SYM_JUMP_IF_FALSE,
                 &&TARGET_CALL_SYMBOL,
                 &&TARGET_GET_FIELD_FROM_SYMBOL,
                 &&TARGET_GET_FIELD_FROM_SYMBOL_INDEX,
@@ -1077,14 +1081,14 @@ namespace Ark
                     TARGET(GT)
                     {
                         const Value *b = popAndResolveAsPtr(context), *a = popAndResolveAsPtr(context);
-                        push((*a != *b && !(*a < *b)) ? Builtins::trueSym : Builtins::falseSym, context);
+                        push(*b < *a ? Builtins::trueSym : Builtins::falseSym, context);
                         DISPATCH();
                     }
 
                     TARGET(LT)
                     {
                         const Value *b = popAndResolveAsPtr(context), *a = popAndResolveAsPtr(context);
-                        push((*a < *b) ? Builtins::trueSym : Builtins::falseSym, context);
+                        push(*a < *b ? Builtins::trueSym : Builtins::falseSym, context);
                         DISPATCH();
                     }
 
@@ -1105,14 +1109,14 @@ namespace Ark
                     TARGET(NEQ)
                     {
                         const Value *b = popAndResolveAsPtr(context), *a = popAndResolveAsPtr(context);
-                        push((*a != *b) ? Builtins::trueSym : Builtins::falseSym, context);
+                        push(*a != *b ? Builtins::trueSym : Builtins::falseSym, context);
                         DISPATCH();
                     }
 
                     TARGET(EQ)
                     {
                         const Value *b = popAndResolveAsPtr(context), *a = popAndResolveAsPtr(context);
-                        push((*a == *b) ? Builtins::trueSym : Builtins::falseSym, context);
+                        push(*a == *b ? Builtins::trueSym : Builtins::falseSym, context);
                         DISPATCH();
                     }
 
@@ -1577,6 +1581,36 @@ namespace Ark
                         DISPATCH();
                     }
 
+                    TARGET(GT_CONST_JUMP_IF_TRUE)
+                    {
+                        UNPACK_ARGS();
+                        const Value* sym = popAndResolveAsPtr(context);
+                        const Value* cst = loadConstAsPtr(primary_arg);
+                        if (*cst < *sym)
+                            context.ip = secondary_arg * 4;
+                        DISPATCH();
+                    }
+
+                    TARGET(GT_CONST_JUMP_IF_FALSE)
+                    {
+                        UNPACK_ARGS();
+                        const Value* sym = popAndResolveAsPtr(context);
+                        const Value* cst = loadConstAsPtr(primary_arg);
+                        if (!(*cst < *sym))
+                            context.ip = secondary_arg * 4;
+                        DISPATCH();
+                    }
+
+                    TARGET(GT_SYM_JUMP_IF_FALSE)
+                    {
+                        UNPACK_ARGS();
+                        const Value* sym = popAndResolveAsPtr(context);
+                        const Value* rhs = loadSymbol(primary_arg, context);
+                        if (!(*rhs < *sym))
+                            context.ip = secondary_arg * 4;
+                        DISPATCH();
+                    }
+
                     TARGET(EQ_CONST_JUMP_IF_TRUE)
                     {
                         UNPACK_ARGS();
@@ -1600,6 +1634,15 @@ namespace Ark
                         UNPACK_ARGS();
                         const Value* sym = popAndResolveAsPtr(context);
                         if (*sym != *loadConstAsPtr(primary_arg))
+                            context.ip = secondary_arg * 4;
+                        DISPATCH();
+                    }
+
+                    TARGET(NEQ_SYM_JUMP_IF_FALSE)
+                    {
+                        UNPACK_ARGS();
+                        const Value* sym = popAndResolveAsPtr(context);
+                        if (*sym == *loadSymbol(primary_arg, context))
                             context.ip = secondary_arg * 4;
                         DISPATCH();
                     }